# openapi.yaml
openapi: 3.1.0
info:
  description: LensBridge backend — media upload and MusallahBoard management
  title: LensBridge API
  version: 0.0.1-SNAPSHOT
servers:
- url: /
security:
- bearerAuth: []
paths:
...
  /api/agent/enroll:
    post:
      description: Called once per device by the MusallahBoard agent. The returned
        websocketUrl is persisted verbatim into the agent's config and never requested
        again.
      operationId: enrollAgent
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentEnrollRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentEnrollResponse"
          description: Device enrolled; identity and websocket URL returned
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Public key malformed or unsupported
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: "Enrollment token invalid, expired, or already used"
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
          description: Request failed; body carries a human-readable message
      summary: Exchange a one-time enrollment token for a device identity
      tags:
      - agent-enrollment-controller
...
components:
  schemas:
...
    AgentEnrollRequest:
      type: object
      properties:
        agentVersion:
          type: string
        hardwareModel:
          type: string
        hostname:
          type: string
          minLength: 1
        publicKey:
          type: string
          minLength: 1
        token:
          type: string
          minLength: 1
      required:
      - hostname
      - publicKey
      - token
    AgentEnrollResponse:
      type: object
      properties:
        deviceId:
          type: string
          format: uuid
        websocketUrl:
          type: string
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      name: bearerAuth
      scheme: bearer
      type: http